CPR-489 add UUID to LIBRA cases
